requirements.txt
aniso8601==7.0.0
asgiref==3.4.1
attrs==21.2.0
Django==3.1.14
django-filter==21.1
django-graphql-auth==0.3.7
django-graphql-jwt==0.3.4
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
promise==2.3
py==1.11.0
PyJWT==2.3.0
pyparsing==3.0.6
pytest==6.2.5
pytest-django==4.5.2
pytz==2021.3
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2